move VersionStatus classes out of semantic_version
