+++ tests/platform_helper/test_command_generate.py
@@ -6,8 +6,8 @@ from click.testing import CliRunner
 
 from dbt_platform_helper.commands.generate import generate as platform_helper_generate
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
-from dbt_platform_helper.providers.semantic_version import PlatformHelperVersionStatus
 from dbt_platform_helper.providers.semantic_version import SemanticVersion
+from dbt_platform_helper.providers.version_status import PlatformHelperVersionStatus
 
 
 @patch("dbt_platform_helper.commands.copilot.get_aws_session_or_abort")
